big merge

+++ dao-orm/src/main/java/com/training/carsharing/repository/AdRepository.java
@@ -2,5 +2,5 @@ package com.training.carsharing.repository;
 
 import com.training.carsharing.model.impl.Ad;
 
-public interface AdRepository extends AbstractRepository<Ad, Integer> {
+public interface AdRepository extends AbstractRepository<Ad, Long> {
 }
