+++ dao-orm/src/main/java/com/training/carsharing/repository/BodyTypeRepository.java
@@ -2,6 +2,6 @@ package com.training.carsharing.repository;
 
 import com.training.carsharing.model.impl.BodyType;
 
-public interface BodyTypeRepository extends AbstractRepository<BodyType, Integer> {
+public interface BodyTypeRepository extends AbstractRepository<BodyType, Long> {
 
 }
