+++ dao-orm/src/main/java/com/training/carsharing/repository/BrandRepository.java
@@ -1,9 +1,8 @@
 package com.training.carsharing.repository;
 
 import com.training.carsharing.model.impl.Brand;
-import org.springframework.stereotype.Repository;
 
 //@Repository
-public interface BrandRepository extends AbstractRepository<Brand, Integer> {
+public interface BrandRepository extends AbstractRepository<Brand, Long> {
 
 }
