+++ dao-orm/src/main/java/com/training/carsharing/repository/CalendarRepository.java
@@ -2,5 +2,5 @@ package com.training.carsharing.repository;
 
 import com.training.carsharing.model.impl.Calendar;
 
-public interface CalendarRepository extends AbstractRepository<Calendar, Integer> {
+public interface CalendarRepository extends AbstractRepository<Calendar, Long> {
 }
