+++ dao-orm/src/main/java/com/training/carsharing/repository/CarRepository.java
@@ -2,5 +2,5 @@ package com.training.carsharing.repository;
 
 import com.training.carsharing.model.impl.Car;
 
-public interface CarRepository extends AbstractRepository<Car, Integer> {
+public interface CarRepository extends AbstractRepository<Car, Long> {
 }
