+++ dao-orm/src/main/java/com/training/carsharing/repository/DrivingLicenseRepository.java
@@ -2,6 +2,6 @@ package com.training.carsharing.repository;
 
 import com.training.carsharing.model.impl.DrivingLicense;
 
-public interface DrivingLicenseRepository extends AbstractRepository<DrivingLicense, Integer> {
+public interface DrivingLicenseRepository extends AbstractRepository<DrivingLicense, Long> {
 
 }
