+++ dao-orm/src/main/java/com/training/carsharing/repository/EngineTypeRepository.java
@@ -2,6 +2,6 @@ package com.training.carsharing.repository;
 
 import com.training.carsharing.model.impl.EngineType;
 
-public interface EngineTypeRepository extends AbstractRepository<EngineType, Integer> {
+public interface EngineTypeRepository extends AbstractRepository<EngineType, Long> {
 
 }
