+++ dao-orm/src/main/java/com/training/carsharing/repository/MessageRepository.java
@@ -2,5 +2,5 @@ package com.training.carsharing.repository;
 
 import com.training.carsharing.model.impl.Message;
 
-public interface MessageRepository extends AbstractRepository<Message, Integer> {
+public interface MessageRepository extends AbstractRepository<Message, Long> {
 }
