+++ dao-orm/src/main/java/com/training/carsharing/repository/PassportRepository.java
@@ -2,6 +2,6 @@ package com.training.carsharing.repository;
 
 import com.training.carsharing.model.impl.Passport;
 
-public interface PassportRepository extends AbstractRepository<Passport, Integer> {
+public interface PassportRepository extends AbstractRepository<Passport, Long> {
 
 }
