+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/BodyTypeRepositoryCustom.java
@@ -1,8 +1,7 @@
 package com.training.carsharing.repository.customrepository;
 
 import com.training.carsharing.model.impl.BodyType;
-import com.training.carsharing.repository.AbstractRepository;
 
-public interface BodyTypeRepositoryCustom extends AbstractRepositoryCustom<BodyType, Integer> {
+public interface BodyTypeRepositoryCustom extends AbstractRepositoryCustom<BodyType, Long> {
 
 }
