+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/BrandRepositoryCustom.java
@@ -1,8 +1,7 @@
 package com.training.carsharing.repository.customrepository;
 
 import com.training.carsharing.model.impl.Brand;
-import com.training.carsharing.repository.AbstractRepository;
 
-public interface BrandRepositoryCustom extends AbstractRepositoryCustom<Brand, Integer> {
+public interface BrandRepositoryCustom extends AbstractRepositoryCustom<Brand, Long> {
 
 }
