+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/CarParameterRepositoryCustom.java
@@ -1,8 +1,7 @@
 package com.training.carsharing.repository.customrepository;
 
 import com.training.carsharing.model.impl.CarParameter;
-import com.training.carsharing.repository.AbstractRepository;
 
-public interface CarParameterRepositoryCustom extends AbstractRepositoryCustom<CarParameter, Integer> {
+public interface CarParameterRepositoryCustom extends AbstractRepositoryCustom<CarParameter, Long> {
 
 }
