+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/DriveRepositoryCustom.java
@@ -1,8 +1,7 @@
 package com.training.carsharing.repository.customrepository;
 
 import com.training.carsharing.model.impl.Drive;
-import com.training.carsharing.repository.AbstractRepository;
 
-public interface DriveRepositoryCustom extends AbstractRepositoryCustom<Drive, Integer> {
+public interface DriveRepositoryCustom extends AbstractRepositoryCustom<Drive, Long> {
 
 }
