+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/DrivingLicenseRepositoryCustom.java
@@ -1,8 +1,7 @@
 package com.training.carsharing.repository.customrepository;
 
 import com.training.carsharing.model.impl.DrivingLicense;
-import com.training.carsharing.repository.AbstractRepository;
 
-public interface DrivingLicenseRepositoryCustom extends AbstractRepositoryCustom<DrivingLicense, Integer> {
+public interface DrivingLicenseRepositoryCustom extends AbstractRepositoryCustom<DrivingLicense, Long> {
 
 }
