+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/EngineTypeRepositoryCustom.java
@@ -1,8 +1,7 @@
 package com.training.carsharing.repository.customrepository;
 
 import com.training.carsharing.model.impl.EngineType;
-import com.training.carsharing.repository.AbstractRepository;
 
-public interface EngineTypeRepositoryCustom extends AbstractRepositoryCustom<EngineType, Integer> {
+public interface EngineTypeRepositoryCustom extends AbstractRepositoryCustom<EngineType, Long> {
 
 }
