+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/FuelRepositoryCustom.java
@@ -1,8 +1,7 @@
 package com.training.carsharing.repository.customrepository;
 
 import com.training.carsharing.model.impl.Fuel;
-import com.training.carsharing.repository.AbstractRepository;
 
-public interface FuelRepositoryCustom extends AbstractRepositoryCustom<Fuel, Integer> {
+public interface FuelRepositoryCustom extends AbstractRepositoryCustom<Fuel, Long> {
 
 }
