+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/GearboxRepositoryCustom.java
@@ -1,8 +1,7 @@
 package com.training.carsharing.repository.customrepository;
 
 import com.training.carsharing.model.impl.Gearbox;
-import com.training.carsharing.repository.AbstractRepository;
 
-public interface GearboxRepositoryCustom extends AbstractRepositoryCustom<Gearbox, Integer> {
+public interface GearboxRepositoryCustom extends AbstractRepositoryCustom<Gearbox, Long> {
 
 }
