+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/MessageRepositoryCustom.java
@@ -1,7 +1,6 @@
 package com.training.carsharing.repository.customrepository;
 
 import com.training.carsharing.model.impl.Message;
-import com.training.carsharing.repository.AbstractRepository;
 
-public interface MessageRepositoryCustom extends AbstractRepositoryCustom<Message, Integer> {
+public interface MessageRepositoryCustom extends AbstractRepositoryCustom<Message, Long> {
 }
