+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/ModelRepositoryCustom.java
@@ -1,8 +1,7 @@
 package com.training.carsharing.repository.customrepository;
 
 import com.training.carsharing.model.impl.Model;
-import com.training.carsharing.repository.AbstractRepository;
 
-public interface ModelRepositoryCustom extends AbstractRepositoryCustom<Model, Integer> {
+public interface ModelRepositoryCustom extends AbstractRepositoryCustom<Model, Long> {
 
 }
