+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/PassportRepositoryCustom.java
@@ -1,8 +1,7 @@
 package com.training.carsharing.repository.customrepository;
 
 import com.training.carsharing.model.impl.Passport;
-import com.training.carsharing.repository.AbstractRepository;
 
-public interface PassportRepositoryCustom extends AbstractRepositoryCustom<Passport, Integer> {
+public interface PassportRepositoryCustom extends AbstractRepositoryCustom<Passport, Long> {
 
 }
